rename types to data
